+++ src/main/java/insane96mcp/shieldsplus/mixin/MinecraftMixin.java
@@ -23,16 +23,40 @@ public abstract class MinecraftMixin {
 
     @Shadow protected abstract boolean startAttack();
 
+    @Shadow protected abstract void startUseItem();
+
     @Definition(id = "options", field = "Lnet/minecraft/client/Minecraft;options:Lnet/minecraft/client/Options;")
     @Definition(id = "keyAttack", field = "Lnet/minecraft/client/Options;keyAttack:Lnet/minecraft/client/KeyMapping;")
     @Definition(id = "consumeClick", method = "Lnet/minecraft/client/KeyMapping;consumeClick()Z")
     @Expression("this.options.keyAttack.consumeClick()")
     @ModifyExpressionValue(method = "handleKeybinds", at = @At(value = "MIXINEXTRAS:EXPRESSION", ordinal = 0))
-    public boolean shieldsPlus$onConsumeAttackClick(boolean original, @Local LocalBooleanRef flag2) {
+    public boolean shieldsPlus$onConsumeAttackClickWhenUsingItem(boolean original, @Local LocalBooleanRef flag2) {
         if (original && BaseFeature.blockWithCrouch(this.player)) {
             //this.gameMode.releaseUsingItem(this.player);
             flag2.set(flag2.get() | this.startAttack());
         }
         return original;
+    }
+
+    @Definition(id = "options", field = "Lnet/minecraft/client/Minecraft;options:Lnet/minecraft/client/Options;")
+    @Definition(id = "keyUse", field = "Lnet/minecraft/client/Options;keyUse:Lnet/minecraft/client/KeyMapping;")
+    @Definition(id = "consumeClick", method = "Lnet/minecraft/client/KeyMapping;consumeClick()Z")
+    @Expression("this.options.keyUse.consumeClick()")
+    @ModifyExpressionValue(method = "handleKeybinds", at = @At(value = "MIXINEXTRAS:EXPRESSION", ordinal = 0))
+    public boolean shieldsPlus$onConsumeUseClickWhenUsingItem(boolean original, @Local LocalBooleanRef flag2) {
+        if (original && BaseFeature.blockWithCrouch(this.player)) {
+            this.startUseItem();
+        }
+        return original;
+    }
+
+    @Definition(id = "player", field = "Lnet/minecraft/client/Minecraft;player:Lnet/minecraft/client/player/LocalPlayer;")
+    @Definition(id = "isUsingItem", method = "Lnet/minecraft/client/player/LocalPlayer;isUsingItem()Z")
+    @Expression("this.player.isUsingItem()")
+    @ModifyExpressionValue(method = "continueAttack", at = @At(value = "MIXINEXTRAS:EXPRESSION"))
+    public boolean shieldsPlus$onUsingItemCheckWhenContinueAttack(boolean original) {
+        if (!original)
+            return false;
+        return !BaseFeature.blockWithCrouch(this.player);
     }
 }